Extract DI container logic into new `Container` class
